+++ lib/src/nesting.dart
@@ -76,6 +76,13 @@ class NestingLevel {
 
     // Unlink the unused parent from the chain.
     _parent = _parent._parent;
+
+    // TODO(rnystrom): This should walk the entire parent chain looking for
+    // unused levels. Stopping after the first can leave some unused levels on
+    // the stack. This isn't fatal, but it makes the splitter slower.
+    //
+    // However, fixing this causes regression 144 to fail. Investigate what's
+    // going on there and fix that.
   }
 
   String toString() => depth.toString();